WIP to update the models we compare
