--- examples/models/mlflow_server_ab_test_ambassador/conda.yaml
@@ -1,8 +1,0 @@
-name: mlflow-talk
-channels:
-  - defaults
-dependencies:
-  - python=3.6
-  - scikit-learn=0.19.1
-  - pip:
-    - mlflow>=1.0
