+++ examples/models/mlflow_server_ab_test_ambassador/model-a/MLproject
@@ -7,4 +7,4 @@ entry_points:
     parameters:
       alpha: float
       l1_ratio: { type: float, default: 0.1 }
-    command: "python train.py {alpha} {l1_ratio}"
+    command: python train.py --alpha {alpha} --l1_ratio {l1_ratio}
