+++ examples/models/mlflow_server_ab_test_ambassador/model-a/conda.yaml
@@ -4,5 +4,6 @@ channels:
 dependencies:
   - python=3.6
   - pyspark=2.4.0
+  - scikit-learn=0.22.1
   - pip:
       - mlflow>=1.0
